+++ src/main/kotlin/ru/barsky/repository/CountryRepository.kt
@@ -4,7 +4,7 @@ import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.CrudRepository
 import ru.barsky.entity.CountryEntity
 
-interface CountryRepository : CrudRepository<CountryEntity, Long> {
+interface CountryRepository : CrudRepository<CountryEntity, Int> {
 
     fun findByOrderByName(pageable: Pageable): List<CountryEntity>
