+++ src/main/kotlin/ru/barsky/service/CountryService.kt
@@ -6,13 +6,13 @@ interface CountryService {
 
     fun getAllCountries(pageIndex: Int): List<CountryDto>
 
-    fun getCountryById(id: Long): CountryDto
+    fun getCountryById(id: Int): CountryDto
 
     fun search(prefix: String): List<CountryDto>
 
-    fun createCountry(countryDto: CountryDto): Long
+    fun createCountry(countryDto: CountryDto): Int
 
-    fun updateCountry(id: Long, countryDto: CountryDto)
+    fun updateCountry(id: Int, countryDto: CountryDto)
 
-    fun deleteCountry(id: Long)
+    fun deleteCountry(id: Int)
 }
